refactor(phase8.1): разбить bot.py на aiogram Router-модули
bot.py: 1 899 строк → 213 строк (только startup + регистрация роутеров)

Новые модули в src/routers/:
- _state.py      — общее изменяемое состояние (dict/set)
- keyboards.py   — константы кнопок + построители клавиатур + тексты
- inspect.py     — InspectFSM («Осмотр улья», ~150 строк)
- fsm_order.py   — OrderFSM 7 шагов + callbacks продуктов/инструкций (~500 строк)
- worker.py      — WorkerAgent очередь сборки (~165 строк)
- bot_admin.py   — admin-команды stats/yt_*/faq/dev/admin-mode (~445 строк)
- user.py        — /start, меню, голос, главный text-handler (~365 строк)

Порядок роутеров (критичен):
admin_router → bot_admin_router → inspect_router →
fsm_order_router → worker_router → user_router (StateFilter=None — последний)

Каждый роутер имеет setup_*() для инжекции bot и агентов из main().
Тесты: импорты обновлены на src.routers.user + src.routers.keyboards.
