+++ src/bot.py
@@ -4,1750 +4,51 @@ import asyncio
 import logging
 from typing import Optional
 
-from aiogram import Bot, Dispatcher, types, F
-from aiogram.filters import CommandStart, Command, StateFilter
-from aiogram.enums import ChatType
-from aiogram.fsm.context import FSMContext
+from aiogram import Bot, Dispatcher, types
 from aiogram.fsm.storage.memory import MemoryStorage
-from aiogram.types import (
-    InlineKeyboardMarkup,
-    InlineKeyboardButton,
-    FSInputFile,
-    ReplyKeyboardMarkup,
-    KeyboardButton,
-    ReplyKeyboardRemove,
+
+from src.config import (
+    TELEGRAM_BOT_TOKEN, BASE_DIR, BEEKEEPER_CHAT_ID, ADMIN_IDS,
+    TG_SOCKS_PROXY, WORKER_CHAT_IDS,
 )
-
-import httpx
-
-from src.config import TELEGRAM_BOT_TOKEN, BASE_DIR, PDFS_DIR, BEEKEEPER_CHAT_ID, ADMIN_CHAT_ID, ADMIN_IDS, ACTIVE_GROUP_IDS, TG_SOCKS_PROXY, DEVBOT_API_URL, WORKER_CHAT_IDS
-from src.phone_utils import validate_phone, format_phone
+from src import config as app_config
 from src.delivery.tracker import OrderTracker
 from src.integrations.uds import UDSClient, UDSPoller
 from src.integram_client import IntegramClient
-from src import config as app_config
-from src.agents.beebot import (
-    INSTRUCTIONS,
-    CATEGORY_LABELS as _CATEGORY_LABELS,
-    is_products_query,
-    get_top_instruction,
-)
-from src.agents.logist import (
-    LogistAgent,
-    OrderFSM,
-    ORDER_TIMEOUT_SECONDS,
-    format_order_summary,
-)
-from src.agents.inspector import (
-    InspectorAgent,
-    InspectFSM,
-    INSPECT_TIMEOUT_SECONDS,
-)
+from src.agents.logist import LogistAgent
+from src.agents.inspector import InspectorAgent
 from src.agents.admin_chat import AdminChatAgent
 from src.orchestrator import Orchestrator
 from src.agents.analyst import AnalystAgent
 from src.admin import router as admin_router, setup_admin
-from src.agents import worker as worker_agent
-from src.llm_client import VOICE_STYLES, DEFAULT_VOICE
 from src.logging_config import setup_logging
 
+# Роутеры из src/routers/
+from src.routers.inspect import router as inspect_router, setup_inspect
+from src.routers.fsm_order import router as fsm_order_router, setup_fsm_order
+from src.routers.worker import router as worker_router, setup_worker
+from src.routers.bot_admin import router as bot_admin_router, setup_bot_admin
+from src.routers.user import router as user_router, setup_user
+import src.routers._state as _state
+
 logger = logging.getLogger(__name__)
-
-# «Голос Улья» — стиль ответов, выбранный пользователем (user_id → style_id)
-_user_styles: dict[int, str] = {}
-
-# Пользователи в режиме «Ассистент пчеловода»
-_admin_mode_users: set[int] = set()
-
-# CRM-клиент для режима работника (устанавливается в main)
-_worker_crm: Optional[IntegramClient] = None
 
 bot = Bot(token=TELEGRAM_BOT_TOKEN)  # может быть переопределён в main() с прокси
 dp = Dispatcher(storage=MemoryStorage())
 
-# Админ-роутер регистрируется первым — его команды имеют приоритет
+# Порядок регистрации роутеров критичен!
+# admin_router — ПЕРВЫМ (приоритет команд /orders, /status, /track и т.д.)
 dp.include_router(admin_router)
-
-orchestrator = Orchestrator()
-logist = LogistAgent(beekeeper_chat_id=BEEKEEPER_CHAT_ID)
-analyst = AnalystAgent(
-    groq_client=orchestrator._groq,
-    groq_model=orchestrator._model,
-)
-inspector = InspectorAgent()  # KB будет подключена в main() после load_kb()
-admin_chat_agent = AdminChatAgent(
-    groq_client=orchestrator._groq,
-    model=orchestrator._model,
-)
-
-# setup_admin() вызывается в main() — один раз, после подключения CRM
-
-# CRM snapshot — кэш заказов с позициями (обновляется в фоне)
-_crm_snapshot: Optional["CrmSnapshot"] = None  # type: ignore[name-defined]
-
-# Хранилище задач таймаута (user_id → asyncio.Task)
-_timeout_lock = asyncio.Lock()
-_timeout_tasks: dict[int, asyncio.Task] = {}
-
-WELCOME_MESSAGE = """Привет! Я бот-помощник Александра Дмитрова — пчеловода и автора блога о продуктах пчеловодства.
-
-Задай мне любой вопрос о:
-- Продуктах пчеловодства (мёд, перга, прополис, пыльца)
-- Рецептах и дозировках
-- Пчеловодстве и здоровье
-
-Просто напиши свой вопрос или выбери раздел ниже 👇"""
-
-HELP_MESSAGE = """Как пользоваться ботом:
-
-В личных сообщениях — просто напиши вопрос.
-
-В группе — обратись ко мне по имени или ответь на моё сообщение:
-• @AleksandrDmitrov_BEEBOT чем полезна перга?
-• Или reply на моё сообщение с вопросом
-
-Примеры вопросов:
-- Как принимать настойку прополиса?
-- Чем полезна перга?
-- Как укрепить иммунитет ребёнка?
-
-/start — начать
-/ask — задать вопрос
-/help — эта справка
-/products — список продуктов с инструкциями
-/order — оформить заказ
-/voice — сменить стиль ответов (Голос Улья)
-/inspect — персональная консультация (Осмотр улья)
-/admin — включить/выключить режим «Ассистент пчеловода» (админ)
-/faq — топ частых вопросов (админ)
-/yt_check — проверить новые видео на канале (админ)
-/yt_update — обновить базу знаний из YouTube (админ)
-/yt_comments — скачать комментарии с ответами автора (админ)
-/stats [запрос] — аналитика продаж (админ)
-/advice — советы пчеловода для консультанта (админ)
-/dev <задача> — отправить задачу в DEVBOT (админ)
-/orders [статус] — список заказов (админ)
-/order <ID> — детали заказа (админ)
-/status <ID> <статус> — сменить статус (админ)
-/track <ID> <трек> — добавить трек-номер (админ)
-/clients — список клиентов (админ)
-/stock — каталог товаров (админ)"""
-
-VOICE_HIVE_MESSAGE = (
-    "🎙️ *Голос Улья* — выбери стиль ответов:\n\n"
-    "Это не разные персонажи — это разные грани одного пчеловода."
-)
-
-ASK_MESSAGE = "Напишите свой вопрос — я отвечу на основе знаний о продуктах пчеловодства 🐝"
-
-PRODUCTS_MESSAGE = "Выбери продукт, чтобы получить PDF с инструкцией:"
-
-VOICE_MESSAGE = "Голосовые сообщения пока не поддерживаю — напиши вопрос текстом, отвечу сразу 🙂"
-
-BOT_USERNAME = "AleksandrDmitrov_BEEBOT"
-
-
-# ---------------------------------------------------------------------------
-# Тексты кнопок главного меню — используются и при построении клавиатуры,
-# и при сопоставлении входящих текстов с обработчиками.
-# ---------------------------------------------------------------------------
-BTN_ASK        = "💬 Спросить"
-BTN_PRODUCTS   = "📦 Продукты"
-BTN_ORDER      = "🛒 Заказать"
-BTN_INSPECT    = "🔍 Осмотр улья"
-BTN_VOICE      = "🎙 Голос Улья"
-BTN_HELP       = "❓ Помощь"
-BTN_STATS      = "📊 Статистика"
-BTN_ADMIN      = "🤖 Ассистент"
-BTN_QUEUE      = "📦 Очередь склада"
-BTN_VIEW_USER  = "👤 Глазами клиента"
-BTN_VIEW_WORK  = "👷 Глазами работника"
-BTN_BACK_ADMIN = "🔙 Режим Админа"
-BTN_REFRESH    = "🔄 Обновить CRM"
-
-ALL_MENU_BTNS = {
-    BTN_ASK, BTN_PRODUCTS, BTN_ORDER, BTN_INSPECT, BTN_VOICE, BTN_HELP,
-    BTN_STATS, BTN_ADMIN, BTN_QUEUE, BTN_VIEW_USER, BTN_VIEW_WORK,
-    BTN_BACK_ADMIN, BTN_REFRESH,
-}
-
-# Текущий вид для каждого администратора: "admin" | "user" | "worker"
-_admin_view_mode: dict[int, str] = {}
-
-
-def _build_main_keyboard(is_admin: bool = False, view: str = "admin") -> ReplyKeyboardMarkup:
-    """Постоянная нижняя клавиатура.
-
-    view="admin"  — полный режим администратора
-    view="user"   — вид глазами клиента (+ кнопка возврата)
-    view="worker" — вид глазами работника (только кнопка возврата)
-    """
-    if view == "user":
-        return ReplyKeyboardMarkup(keyboard=[
-            [KeyboardButton(text=BTN_ASK), KeyboardButton(text=BTN_PRODUCTS), KeyboardButton(text=BTN_ORDER)],
-            [KeyboardButton(text=BTN_INSPECT), KeyboardButton(text=BTN_VOICE), KeyboardButton(text=BTN_HELP)],
-            [KeyboardButton(text=BTN_BACK_ADMIN)],
-        ], resize_keyboard=True)
-
-    if view == "worker":
-        return ReplyKeyboardMarkup(keyboard=[
-            [KeyboardButton(text=BTN_BACK_ADMIN)],
-        ], resize_keyboard=True)
-
-    # view == "admin"
-    rows = [
-        [KeyboardButton(text=BTN_ASK), KeyboardButton(text=BTN_PRODUCTS), KeyboardButton(text=BTN_ORDER)],
-        [KeyboardButton(text=BTN_INSPECT), KeyboardButton(text=BTN_VOICE), KeyboardButton(text=BTN_HELP)],
-        [KeyboardButton(text=BTN_STATS), KeyboardButton(text=BTN_ADMIN), KeyboardButton(text=BTN_QUEUE)],
-        [KeyboardButton(text=BTN_VIEW_USER), KeyboardButton(text=BTN_VIEW_WORK), KeyboardButton(text=BTN_REFRESH)],
-    ]
-    return ReplyKeyboardMarkup(keyboard=rows, resize_keyboard=True)
-
-
-def _build_start_keyboard() -> InlineKeyboardMarkup:
-    return InlineKeyboardMarkup(inline_keyboard=[[
-        InlineKeyboardButton(text="📦 Все продукты", callback_data="show_products"),
-        InlineKeyboardButton(text="❓ Как пользоваться", callback_data="show_help"),
-    ]])
-
-
-def _build_back_to_products_keyboard() -> InlineKeyboardMarkup:
-    return InlineKeyboardMarkup(inline_keyboard=[[
-        InlineKeyboardButton(text="📦 Все продукты", callback_data="show_products"),
-    ]])
-
-
-def _build_products_keyboard() -> InlineKeyboardMarkup:
-    rows = []
-    current_cat = None
-    cat_buttons = []
-
-    for i, (stem, name, fname, cat) in enumerate(INSTRUCTIONS):
-        if not (PDFS_DIR / fname).exists():
-            continue
-        if cat != current_cat:
-            if cat_buttons:
-                rows += [cat_buttons[j:j+2] for j in range(0, len(cat_buttons), 2)]
-            rows.append([InlineKeyboardButton(
-                text=_CATEGORY_LABELS.get(cat, cat),
-                callback_data="noop",
-            )])
-            cat_buttons = []
-            current_cat = cat
-        cat_buttons.append(InlineKeyboardButton(text=name, callback_data=f"doc:{i}"))
-
-    if cat_buttons:
-        rows += [cat_buttons[j:j+2] for j in range(0, len(cat_buttons), 2)]
-
-    return InlineKeyboardMarkup(inline_keyboard=rows)
-
-
-def _get_instruction_keyboard(chunks: list[dict]) -> InlineKeyboardMarkup:
-    result = get_top_instruction(chunks)
-    if result is None:
-        return InlineKeyboardMarkup(inline_keyboard=[[
-            InlineKeyboardButton(text="📦 Все продукты", callback_data="show_products"),
-        ]])
-
-    idx, name, filename = result
-    if not (PDFS_DIR / filename).exists():
-        return InlineKeyboardMarkup(inline_keyboard=[[
-            InlineKeyboardButton(text="📦 Все продукты", callback_data="show_products"),
-        ]])
-
-    return InlineKeyboardMarkup(inline_keyboard=[[
-        InlineKeyboardButton(text=f"📄 {name}", callback_data=f"doc:{idx}"),
-        InlineKeyboardButton(text="📦 Все продукты", callback_data="show_products"),
-    ]])
-
-
-@dp.message(CommandStart())
-async def cmd_start(message: types.Message):
-    uid = message.from_user.id
-    is_admin = bool(ADMIN_IDS and uid in ADMIN_IDS)
-
-    # Работники склада (не-администраторы) сразу видят очередь сборки
-    if WORKER_CHAT_IDS and uid in WORKER_CHAT_IDS and not is_admin:
-        await _worker_show_queue(message.chat.id, message)
-        return
-
-    # Сбросить вид на "admin" при /start
-    if is_admin:
-        _admin_view_mode[uid] = "admin"
-
-    await message.answer(
-        WELCOME_MESSAGE,
-        reply_markup=_build_main_keyboard(is_admin=is_admin, view="admin"),
-    )
-
-
-# ---------------------------------------------------------------------------
-# Обработчики кнопок главного меню (ReplyKeyboard)
-# ---------------------------------------------------------------------------
-
-@dp.message(F.text == BTN_ASK)
-async def btn_ask(message: types.Message):
-    await message.answer(ASK_MESSAGE)
-
-
-@dp.message(F.text == BTN_PRODUCTS)
-async def btn_products(message: types.Message):
-    await message.answer(PRODUCTS_MESSAGE, reply_markup=_build_products_keyboard())
-
-
-@dp.message(F.text == BTN_ORDER)
-async def btn_order(message: types.Message, state: FSMContext):
-    await cmd_order(message, state)
-
-
-@dp.message(F.text == BTN_INSPECT)
-async def btn_inspect(message: types.Message, state: FSMContext):
-    await cmd_inspect(message, state)
-
-
-@dp.message(F.text == BTN_VOICE)
-async def btn_voice(message: types.Message):
-    current = _user_styles.get(message.from_user.id, DEFAULT_VOICE)
-    await message.answer(
-        VOICE_HIVE_MESSAGE,
-        parse_mode="Markdown",
-        reply_markup=_build_voice_keyboard(current),
-    )
-
-
-@dp.message(F.text == BTN_HELP)
-async def btn_help(message: types.Message):
-    await message.answer(HELP_MESSAGE)
-
-
-@dp.message(F.text == BTN_STATS)
-async def btn_stats(message: types.Message):
-    if not ADMIN_IDS or message.from_user.id not in ADMIN_IDS:
-        return
-    await cmd_stats(message)
-
-
-@dp.message(F.text == BTN_ADMIN)
-async def btn_admin_mode(message: types.Message):
-    if not ADMIN_IDS or message.from_user.id not in ADMIN_IDS:
-        return
-    await cmd_admin_mode(message)
-
-
-@dp.message(F.text == BTN_QUEUE)
-async def btn_queue(message: types.Message):
-    if not ADMIN_IDS or message.from_user.id not in ADMIN_IDS:
-        return
-    await _worker_show_queue(message.chat.id, message)
-
-
-@dp.message(F.text == BTN_VIEW_USER)
-async def btn_view_as_user(message: types.Message):
-    """Администратор переключается в вид «Глазами клиента»."""
-    uid = message.from_user.id
-    if not ADMIN_IDS or uid not in ADMIN_IDS:
-        return
-    _admin_view_mode[uid] = "user"
-    await message.answer(
-        "👤 *Вид «Глазами клиента»*\n\nТеперь ты видишь то, что видят покупатели.\nКнопка «🔙 Режим Админа» вернёт тебя обратно.",
-        parse_mode="Markdown",
-        reply_markup=_build_main_keyboard(is_admin=True, view="user"),
-    )
-
-
-@dp.message(F.text == BTN_VIEW_WORK)
-async def btn_view_as_worker(message: types.Message):
-    """Администратор переключается в вид «Глазами работника»."""
-    uid = message.from_user.id
-    if not ADMIN_IDS or uid not in ADMIN_IDS:
-        return
-    _admin_view_mode[uid] = "worker"
-    await message.answer(
-        "👷 *Вид «Глазами работника»*\n\nПоказываю очередь сборки — как её видит работник склада.",
-        parse_mode="Markdown",
-        reply_markup=_build_main_keyboard(is_admin=True, view="worker"),
-    )
-    await _worker_show_queue(message.chat.id, message)
-
-
-@dp.message(F.text == BTN_BACK_ADMIN)
-async def btn_back_to_admin(message: types.Message):
-    """Вернуться в полный режим администратора."""
-    uid = message.from_user.id
-    if not ADMIN_IDS or uid not in ADMIN_IDS:
-        return
-    _admin_view_mode[uid] = "admin"
-    await message.answer(
-        "🔙 *Режим Админа*",
-        parse_mode="Markdown",
-        reply_markup=_build_main_keyboard(is_admin=True, view="admin"),
-    )
-
-
-@dp.message(F.text == BTN_REFRESH)
-@dp.message(Command("refresh_crm"))
-async def btn_refresh_crm(message: types.Message):
-    """Принудительное обновление CRM snapshot (только админ)."""
-    if not ADMIN_IDS or message.from_user.id not in ADMIN_IDS:
-        return
-    if not _crm_snapshot:
-        await message.answer("⚠️ CRM snapshot не инициализирован.")
-        return
-    msg = await message.answer("🔄 Обновляю снимок CRM...")
-    try:
-        await _crm_snapshot.refresh()
-        await msg.edit_text(
-            f"✅ Снимок CRM обновлён\n"
-            f"📦 Заказов: {len(_crm_snapshot.orders)}\n"
-            f"👥 Клиентов: {len(_crm_snapshot.clients)}\n"
-            f"🍯 Товаров: {len(_crm_snapshot.products)}\n"
-            f"⏱ {_crm_snapshot.age_str}"
-        )
-    except Exception as e:
-        await msg.edit_text(f"❌ Ошибка обновления: {e}")
-
-
-@dp.message(Command("queue"))
-async def cmd_worker_queue(message: types.Message):
-    """Показать очередь сборки (только для работников склада)."""
-    uid = message.from_user.id
-    is_admin = bool(ADMIN_IDS and uid in ADMIN_IDS)
-    if not is_admin and (not WORKER_CHAT_IDS or uid not in WORKER_CHAT_IDS):
-        return
-    await _worker_show_queue(message.chat.id, message)
-
-
-async def _worker_show_queue(chat_id: int, source: types.Message | types.CallbackQuery) -> None:
-    """Отправить/обновить сообщение с очередью заказов."""
-    if not _worker_crm:
-        text = "⚠️ CRM недоступна — попробуйте позже."
-        if isinstance(source, types.Message):
-            await source.answer(text)
-        else:
-            await source.message.edit_text(text)
-        return
-    try:
-        orders = await worker_agent.get_worker_queue(_worker_crm)
-        text = worker_agent.format_queue_text(orders)
-        keyboard = worker_agent.build_queue_keyboard(orders)
-        if isinstance(source, types.Message):
-            await source.answer(text, parse_mode="Markdown", reply_markup=keyboard)
-        else:
-            await source.message.edit_text(text, parse_mode="Markdown", reply_markup=keyboard)
-    except Exception as e:
-        logger.error("Ошибка загрузки очереди работника: %s", e)
-        err = "❌ Не удалось загрузить очередь. Попробуйте ещё раз."
-        if isinstance(source, types.Message):
-            await source.answer(err)
-        else:
-            await source.message.edit_text(err)
-
-
-@dp.callback_query(F.data == "worker:queue")
-async def cb_worker_queue(callback: types.CallbackQuery):
-    if not WORKER_CHAT_IDS or callback.from_user.id not in WORKER_CHAT_IDS:
-        await callback.answer("⛔ Нет доступа.", show_alert=True)
-        return
-    await callback.answer()
-    await _worker_show_queue(callback.message.chat.id, callback)
-
-
-@dp.callback_query(F.data.startswith("worker:order:"))
-async def cb_worker_order(callback: types.CallbackQuery):
-    if not WORKER_CHAT_IDS or callback.from_user.id not in WORKER_CHAT_IDS:
-        await callback.answer("⛔ Нет доступа.", show_alert=True)
-        return
-    await callback.answer()
-    order_id = int(callback.data.split(":")[2])
-    await _worker_show_order(callback, order_id)
-
-
-@dp.callback_query(F.data.startswith("worker:take:"))
-async def cb_worker_take(callback: types.CallbackQuery):
-    """Работник берёт заказ в работу → статус «В сборке»."""
-    if not WORKER_CHAT_IDS or callback.from_user.id not in WORKER_CHAT_IDS:
-        await callback.answer("⛔ Нет доступа.", show_alert=True)
-        return
-    order_id = int(callback.data.split(":")[2])
-    if not _worker_crm:
-        await callback.answer("⚠️ CRM недоступна.", show_alert=True)
-        return
-    try:
-        order = await _worker_crm.get_order(order_id)
-        if order.status not in ("Новый", "Подтверждён"):
-            await callback.answer(f"Заказ уже в статусе «{order.status}».", show_alert=True)
-            return
-        await _worker_crm.update_order_status(order_id, "В сборке", comment="Взят в работу работником склада")
-        await callback.answer("✅ Взят в работу!")
-    except Exception as e:
-        logger.error("Ошибка взятия заказа в работу: %s", e)
-        await callback.answer("❌ Ошибка. Попробуйте ещё раз.", show_alert=True)
-        return
-    await _worker_show_order(callback, order_id)
-
-
-@dp.callback_query(F.data.startswith("worker:check:"))
-async def cb_worker_check(callback: types.CallbackQuery):
-    """Отметить/снять позицию в чеклисте."""
-    if not WORKER_CHAT_IDS or callback.from_user.id not in WORKER_CHAT_IDS:
-        await callback.answer("⛔ Нет доступа.", show_alert=True)
-        return
-    parts = callback.data.split(":")
-    order_id = int(parts[2])
-    item_id = int(parts[3])
-    worker_agent.toggle_item(callback.from_user.id, order_id, item_id)
-    await callback.answer()
-    await _worker_show_order(callback, order_id)
-
-
-@dp.callback_query(F.data.startswith("worker:done:"))
-async def cb_worker_done(callback: types.CallbackQuery):
-    """Работник завершил сборку → уведомить пчеловода."""
-    if not WORKER_CHAT_IDS or callback.from_user.id not in WORKER_CHAT_IDS:
-        await callback.answer("⛔ Нет доступа.", show_alert=True)
-        return
-    order_id = int(callback.data.split(":")[2])
-    if not _worker_crm:
-        await callback.answer("⚠️ CRM недоступна.", show_alert=True)
-        return
-    try:
-        order = await _worker_crm.get_order(order_id)
-        # Отметить «Наличие проверено» в чеклисте CRM
-        await _worker_crm.update_order_checklist(order_id, stock_checked=True)
-        worker_agent.clear_checklist(callback.from_user.id, order_id)
-
-        # Уведомить пчеловода
-        from src.notifications import _worker_notifier
-        if _worker_notifier:
-            await _worker_notifier.notify_workers_assembled(order_id, order.number or str(order_id))
-
-        await callback.answer("📦 Пчеловод уведомлён!", show_alert=True)
-    except Exception as e:
-        logger.error("Ошибка завершения сборки: %s", e)
-        await callback.answer("❌ Ошибка. Попробуйте ещё раз.", show_alert=True)
-        return
-
-    # Вернуться к очереди
-    await _worker_show_queue(callback.message.chat.id, callback)
-
-
-async def _worker_show_order(callback: types.CallbackQuery, order_id: int) -> None:
-    """Показать карточку заказа с чеклистом."""
-    if not _worker_crm:
-        await callback.message.edit_text("⚠️ CRM недоступна.")
-        return
-    try:
-        order = await _worker_crm.get_order(order_id)
-        items = await _worker_crm.get_order_items(order_id)
-        text = worker_agent.format_order_card(order, items, callback.from_user.id)
-        keyboard = worker_agent.build_order_keyboard(order_id, items, callback.from_user.id, order.status)
-        await callback.message.edit_text(text, parse_mode="Markdown", reply_markup=keyboard)
-    except Exception as e:
-        logger.error("Ошибка загрузки карточки заказа %d: %s", order_id, e)
-        await callback.message.edit_text("❌ Не удалось загрузить заказ. Попробуйте ещё раз.")
-
-
-@dp.message(Command("help"))
-async def cmd_help(message: types.Message):
-    await message.answer(HELP_MESSAGE, reply_markup=_build_back_to_products_keyboard())
-
-
-@dp.message(Command("ask"))
-async def cmd_ask(message: types.Message):
-    await message.answer(ASK_MESSAGE)
-
-
-@dp.message(Command("products"))
-async def cmd_products(message: types.Message):
-    await message.answer(PRODUCTS_MESSAGE, reply_markup=_build_products_keyboard())
-
-
-def _build_voice_keyboard(current_style: str) -> InlineKeyboardMarkup:
-    """Клавиатура выбора «Голоса Улья»."""
-    rows = []
-    for sid, sdata in VOICE_STYLES.items():
-        check = "✅ " if sid == current_style else ""
-        label = f"{check}{sdata['emoji']} {sdata['name']} — {sdata['desc']}"
-        rows.append([InlineKeyboardButton(text=label, callback_data=f"voice_style:{sid}")])
-    return InlineKeyboardMarkup(inline_keyboard=rows)
-
-
-@dp.message(Command("voice"))
-async def cmd_voice(message: types.Message):
-    """Выбрать стиль ответов «Голос Улья»."""
-    current = _user_styles.get(message.from_user.id, DEFAULT_VOICE)
-    await message.answer(
-        VOICE_HIVE_MESSAGE,
-        parse_mode="Markdown",
-        reply_markup=_build_voice_keyboard(current),
-    )
-
-
-@dp.message(Command("admin"))
-async def cmd_admin_mode(message: types.Message):
-    """Переключить режим «Ассистент пчеловода» (только ADMIN_CHAT_ID)."""
-    if not ADMIN_IDS or message.from_user.id not in ADMIN_IDS:
-        await message.answer("⛔ Команда доступна только администратору.")
-        return
-
-    user_id = message.from_user.id
-    if user_id in _admin_mode_users:
-        _admin_mode_users.discard(user_id)
-        admin_chat_agent.clear_history(user_id)
-        await message.answer(
-            "🐝 Режим *Ассистент* выключен.\n"
-            "Снова работаю как бот для подписчиков.",
-            parse_mode="Markdown",
-        )
-    else:
-        _admin_mode_users.add(user_id)
-        await message.answer(
-            "🤖 *Режим: Ассистент пчеловода*\n\n"
-            "Теперь я твой личный помощник с доступом к CRM.\n"
-            "Спрашивай что угодно:\n\n"
-            "• _Сколько заказов за этот месяц?_\n"
-            "• _Какие товары нужно срочно пополнить?_\n"
-            "• _Напиши ответ клиенту на жалобу о задержке_\n"
-            "• _Топ продуктов за всё время_\n\n"
-            "/admin — выключить режим и очистить историю",
-            parse_mode="Markdown",
-        )
-
-
-@dp.callback_query(F.data.startswith("voice_style:"))
-async def cb_voice_style(callback: types.CallbackQuery):
-    """Сохранить выбранный стиль «Голос Улья»."""
-    style_id = callback.data.split(":", 1)[1]
-    if style_id not in VOICE_STYLES:
-        await callback.answer("Неизвестный стиль", show_alert=True)
-        return
-    _user_styles[callback.from_user.id] = style_id
-    sdata = VOICE_STYLES[style_id]
-    await callback.answer(f"{sdata['emoji']} {sdata['name']} активирован!")
-    await callback.message.edit_text(
-        f"✅ Стиль *{sdata['emoji']} {sdata['name']}* выбран.\n\n"
-        f"_{sdata['desc']}_\n\n"
-        "Задай любой вопрос — отвечу в этом стиле.",
-        parse_mode="Markdown",
-    )
-
-
-# ===========================================================================
-# «Осмотр улья» — диагностический квест
-# ===========================================================================
-
-def _inspect_skip_keyboard() -> InlineKeyboardMarkup:
-    """Кнопка для пропуска оставшихся вопросов."""
-    return InlineKeyboardMarkup(inline_keyboard=[[
-        InlineKeyboardButton(text="💡 Получить рекомендацию", callback_data="inspect_finish"),
-    ]])
-
-
-@dp.message(Command("inspect"))
-async def cmd_inspect(message: types.Message, state: FSMContext):
-    """Запустить диагностический диалог «Осмотр улья»."""
-    current_state = await state.get_state()
-    if current_state and current_state.startswith("OrderFSM"):
-        await message.answer("Сначала заверши или отмени текущий заказ (/cancel).")
-        return
-
-    await state.set_state(InspectFSM.describing_issue)
-    await state.update_data(inspect_collected=[])
-    await message.answer(
-        "🐝 *Осмотр улья* — персональная консультация\n\n"
-        "Опиши, что тебя беспокоит или чего хочешь достичь. "
-        "Я задам пару уточняющих вопросов и дам точную рекомендацию.",
-        parse_mode="Markdown",
-        reply_markup=InlineKeyboardMarkup(inline_keyboard=[[
-            InlineKeyboardButton(text="❌ Отменить", callback_data="inspect_cancel"),
-        ]]),
-    )
-
-
-@dp.message(InspectFSM.describing_issue)
-async def inspect_got_issue(message: types.Message, state: FSMContext):
-    """Получить описание проблемы → задать 1-й уточняющий вопрос."""
-    issue = (message.text or "").strip()
-    if len(issue) < 5:
-        await message.answer("Расскажи подробнее — хотя бы несколько слов.")
-        return
-
-    collected = [issue]
-    await state.update_data(inspect_collected=collected)
-    await bot.send_chat_action(message.chat.id, "typing")
-
-    question = inspector.generate_question(collected)
-    await state.set_state(InspectFSM.answering_q1)
-    await message.answer(question, reply_markup=_inspect_skip_keyboard())
-
-
-@dp.message(InspectFSM.answering_q1)
-async def inspect_got_q1(message: types.Message, state: FSMContext):
-    """Получить ответ на 1-й вопрос → задать 2-й или завершить."""
-    answer = (message.text or "").strip()
-    data = await state.get_data()
-    collected: list[str] = data.get("inspect_collected", [])
-    collected.append(answer)
-    await state.update_data(inspect_collected=collected)
-    await bot.send_chat_action(message.chat.id, "typing")
-
-    question = inspector.generate_question(collected)
-    await state.set_state(InspectFSM.answering_q2)
-    await message.answer(question, reply_markup=_inspect_skip_keyboard())
-
-
-@dp.message(InspectFSM.answering_q2)
-async def inspect_got_q2(message: types.Message, state: FSMContext):
-    """Получить ответ на 2-й вопрос → выдать рекомендацию."""
-    answer = (message.text or "").strip()
-    data = await state.get_data()
-    collected: list[str] = data.get("inspect_collected", [])
-    collected.append(answer)
-    await state.clear()
-
-    await bot.send_chat_action(message.chat.id, "typing")
-    style = _user_styles.get(message.from_user.id)
-    recommendation = inspector.generate_recommendation(collected, style=style)
-    await message.answer(
-        f"🍯 *Рекомендация*\n\n{recommendation}",
-        parse_mode="Markdown",
-        reply_markup=InlineKeyboardMarkup(inline_keyboard=[[
-            InlineKeyboardButton(text="🔄 Новый осмотр", callback_data="inspect_restart"),
-            InlineKeyboardButton(text="📦 Все продукты", callback_data="show_products"),
-        ]]),
-    )
-
-
-@dp.callback_query(F.data == "inspect_finish")
-async def inspect_cb_finish(callback: types.CallbackQuery, state: FSMContext):
-    """Досрочно завершить диалог и получить рекомендацию."""
-    data = await state.get_data()
-    collected: list[str] = data.get("inspect_collected", [])
-    await state.clear()
-
-    if not collected:
-        await callback.answer("Сначала опиши проблему.", show_alert=True)
-        return
-
-    await callback.answer()
-    await callback.message.edit_reply_markup(reply_markup=None)
-    await bot.send_chat_action(callback.message.chat.id, "typing")
-    style = _user_styles.get(callback.from_user.id)
-    recommendation = inspector.generate_recommendation(collected, style=style)
-    await callback.message.answer(
-        f"🍯 *Рекомендация*\n\n{recommendation}",
-        parse_mode="Markdown",
-        reply_markup=InlineKeyboardMarkup(inline_keyboard=[[
-            InlineKeyboardButton(text="🔄 Новый осмотр", callback_data="inspect_restart"),
-            InlineKeyboardButton(text="📦 Все продукты", callback_data="show_products"),
-        ]]),
-    )
-
-
-@dp.callback_query(F.data == "inspect_cancel")
-async def inspect_cb_cancel(callback: types.CallbackQuery, state: FSMContext):
-    """Отменить диагностику."""
-    await state.clear()
-    await callback.answer("Осмотр отменён.")
-    await callback.message.edit_text("Осмотр отменён. Задай вопрос в любое время.")
-
-
-@dp.callback_query(F.data == "inspect_restart")
-async def inspect_cb_restart(callback: types.CallbackQuery, state: FSMContext):
-    """Начать осмотр заново."""
-    await callback.answer()
-    await state.set_state(InspectFSM.describing_issue)
-    await state.update_data(inspect_collected=[])
-    await callback.message.answer(
-        "🐝 *Новый осмотр*\n\nОпиши, что тебя беспокоит.",
-        parse_mode="Markdown",
-        reply_markup=InlineKeyboardMarkup(inline_keyboard=[[
-            InlineKeyboardButton(text="❌ Отменить", callback_data="inspect_cancel"),
-        ]]),
-    )
-
-
-@dp.message(Command("stats"))
-async def cmd_stats(message: types.Message):
-    """Аналитика продаж — только для ADMIN_CHAT_ID."""
-    if not ADMIN_IDS or message.from_user.id not in ADMIN_IDS:
-        await message.answer("⛔ Команда доступна только администратору.")
-        return
-
-    # Получить текст запроса из аргументов команды
-    query = (message.text or "").removeprefix("/stats").strip()
-    if not query:
-        query = "общая статистика"
-
-    await bot.send_chat_action(message.chat.id, "typing")
-    try:
-        report = await analyst.handle_query(query)
-        await message.answer(report, parse_mode="Markdown")
-    except Exception as e:
-        logger.error("Ошибка аналитики: %s", e)
-        await message.answer("Не удалось получить статистику. Попробуйте позже.")
-
-
-# ===========================================================================
-# YouTube updater — /yt_check, /yt_update  (только ADMIN_CHAT_ID)
-# ===========================================================================
-
-@dp.message(Command("yt_check"))
-async def cmd_yt_check(message: types.Message):
-    """Проверить новые видео на YouTube-канале."""
-    if not ADMIN_IDS or message.from_user.id not in ADMIN_IDS:
-        await message.answer("⛔ Команда доступна только администратору.")
-        return
-
-    from src import config as _cfg
-    if not _cfg.YOUTUBE_API_KEY:
-        await message.answer(
-            "⚠️ YouTube API недоступен.\n\n"
-            "Добавь в `.env`:\n"
-            "```\nYOUTUBE_API_KEY=ваш_ключ\n```\n"
-            "Получить ключ: console.cloud.google.com → YouTube Data API v3",
-            parse_mode="Markdown",
-        )
-        return
-
-    await bot.send_chat_action(message.chat.id, "typing")
-    from src.youtube_updater import check_new_videos
-    try:
-        all_ids, new_ids = await check_new_videos(_cfg.YOUTUBE_API_KEY, _cfg.YOUTUBE_CHANNEL_HANDLE)
-        if not all_ids:
-            await message.answer("❌ Не удалось получить список видео. Проверь YOUTUBE_API_KEY.")
-            return
-        from src.youtube_updater import _get_known_ids
-        text = (
-            f"📺 *Канал {_cfg.YOUTUBE_CHANNEL_HANDLE}*\n\n"
-            f"Всего видео: {len(all_ids)}\n"
-            f"Субтитров в KB: {len(_get_known_ids())}\n"
-            f"*Новых: {len(new_ids)}*"
-        )
-        if new_ids:
-            text += f"\n\nID новых видео:\n" + "\n".join(f"• `{v}`" for v in new_ids[:10])
-            text += f"\n\nДля загрузки используй /yt\\_update"
-        else:
-            text += "\n\n✅ База знаний актуальна — новых видео нет."
-        await message.answer(text, parse_mode="Markdown")
-    except Exception as e:
-        logger.error("yt_check error: %s", e)
-        await message.answer(f"Ошибка при проверке: {e}")
-
-
-@dp.message(Command("yt_update"))
-async def cmd_yt_update(message: types.Message):
-    """Скачать субтитры новых видео и пересобрать базу знаний."""
-    if not ADMIN_IDS or message.from_user.id not in ADMIN_IDS:
-        await message.answer("⛔ Команда доступна только администратору.")
-        return
-
-    from src import config as _cfg
-    if not _cfg.YOUTUBE_API_KEY:
-        await message.answer("⚠️ YOUTUBE_API_KEY не задан в .env.")
-        return
-
-    await message.answer("⏳ Проверяю новые видео и обновляю базу знаний...")
-    await bot.send_chat_action(message.chat.id, "typing")
-    from src.youtube_updater import run_update
-    try:
-        report = await run_update(_cfg.YOUTUBE_API_KEY, _cfg.YOUTUBE_CHANNEL_HANDLE)
-        # Перезагрузить KB в боте если пересобрана
-        if "пересобрана" in report:
-            try:
-                orchestrator.load_kb()
-                inspector.kb = orchestrator._beebot.kb
-                report += f"\n\n🔄 KB перезагружена в боте: {len(orchestrator._beebot.kb.chunks)} чанков"
-            except Exception as e:
-                report += f"\n\n⚠️ KB пересобрана, но перезагрузка не удалась: {e}"
-        await message.answer(report, parse_mode="Markdown")
-    except Exception as e:
-        logger.error("yt_update error: %s", e)
-        await message.answer(f"Ошибка при обновлении: {e}")
-
-
-@dp.message(Command("yt_comments"))
-async def cmd_yt_comments(message: types.Message):
-    """Скачать комментарии с ответами автора и добавить в базу знаний."""
-    if not ADMIN_IDS or message.from_user.id not in ADMIN_IDS:
-        await message.answer("⛔ Команда доступна только администратору.")
-        return
-
-    from src import config as _cfg
-    if not _cfg.YOUTUBE_API_KEY:
-        await message.answer("⚠️ YOUTUBE_API_KEY не задан в .env.")
-        return
-
-    await message.answer("⏳ Скачиваю комментарии с ответами автора (~2 мин)...")
-    await bot.send_chat_action(message.chat.id, "typing")
-
-    import asyncio
-    from src.youtube_comments import download_all_comments
-    from src.youtube_loader import CHANNEL_VIDEO_IDS
-    from src.youtube_updater import rebuild_knowledge_base
-
-    try:
-        results = await asyncio.to_thread(
-            download_all_comments,
-            CHANNEL_VIDEO_IDS,
-            _cfg.YOUTUBE_API_KEY,
-            _cfg.YOUTUBE_CHANNEL_HANDLE,
-        )
-        videos_with_qa = sum(1 for n in results.values() if n > 0)
-        total_pairs = sum(results.values())
-
-        report = (
-            f"💬 *Комментарии с ответами автора*\n\n"
-            f"Видео обработано: {len(results)}\n"
-            f"Видео с Q&A: {videos_with_qa}\n"
-            f"Всего Q&A пар: *{total_pairs}*\n\n"
-            f"Пересобираю базу знаний..."
-        )
-        await message.answer(report, parse_mode="Markdown")
-
-        n_chunks = await asyncio.to_thread(rebuild_knowledge_base)
-
-        try:
-            orchestrator.load_kb()
-            inspector.kb = orchestrator._beebot.kb
-            reload_msg = f"🔄 KB перезагружена в боте: {len(orchestrator._beebot.kb.chunks)} чанков"
-        except Exception as e:
-            reload_msg = f"⚠️ KB пересобрана, но перезагрузка не удалась: {e}"
-
-        await message.answer(
-            f"✅ База знаний пересобрана: *{n_chunks} чанков*\n{reload_msg}",
-            parse_mode="Markdown",
-        )
-    except Exception as e:
-        logger.error("yt_comments error: %s", e)
-        await message.answer(f"Ошибка: {e}")
-
-
-# ===========================================================================
-# FAQ — /faq  (только ADMIN_CHAT_ID)
-# ===========================================================================
-
-@dp.message(Command("faq"))
-async def cmd_faq(message: types.Message):
-    """Показать топ частых вопросов пользователей."""
-    if not ADMIN_IDS or message.from_user.id not in ADMIN_IDS:
-        await message.answer("⛔ Команда доступна только администратору.")
-        return
-
-    # Принудительно сохранить FAQ на диск
-    orchestrator.flush_faq()
-
-    # Получить аргумент N (по умолчанию 20)
-    args = (message.text or "").removeprefix("/faq").strip()
-    try:
-        n = int(args) if args else 20
-        n = max(5, min(50, n))
-    except ValueError:
-        n = 20
-
-    top = orchestrator.get_top_queries(n)
-    if not top:
-        await message.answer("📝 FAQ пока пуст — пользователи ещё не задавали вопросов.")
-        return
-
-    total = sum(c for _, c in top)
-    lines = [f"📝 *Топ-{n} частых вопросов* (всего запросов: {total}):\n"]
-    for i, (query, count) in enumerate(top, 1):
-        bar = "▪" * min(count, 10)
-        lines.append(f"{i}. [{count}] {bar} {query}")
-
-    await message.answer("\n".join(lines), parse_mode="Markdown")
-
-
-# ===========================================================================
-# /advice — советы пчеловода (только ADMIN_CHAT_ID)
-# ===========================================================================
-
-@dp.message(Command("advice"))
-async def cmd_advice(message: types.Message):
-    """Показать загруженные советы пчеловода (инжектируются в промпт консультанта)."""
-    if not ADMIN_IDS or message.from_user.id not in ADMIN_IDS:
-        await message.answer("⛔ Команда доступна только администратору.")
-        return
-
-    items = orchestrator._ontology.advice_items
-    if not items:
-        await message.answer(
-            "📭 Советы пчеловода не загружены.\n\n"
-            "Добавь записи в таблицу 7195 (Советы пчеловода) в Integram — "
-            "они подхватятся при следующем перезапуске бота."
-        )
-        return
-
-    _pri_emoji = {"высокий": "🔴", "средний": "🟡", "справочный": "⚪"}
-    lines = [f"🐝 *Советы пчеловода* ({len(items)} активных):\n"]
-    for item in items:
-        pri = _pri_emoji.get(item["priority"], "⚪")
-        cat = f" [{item['category']}]" if item["category"] else ""
-        text = item["text"] or item["name"]
-        lines.append(f"{pri}{cat} {text[:200]}")
-
-    await message.answer("\n".join(lines), parse_mode="Markdown")
-
-
-# ===========================================================================
-# /dev — отправить задачу в DEVBOT (только ADMIN_CHAT_ID)
-# ===========================================================================
-
-@dp.message(Command("dev"))
-async def cmd_dev(message: types.Message):
-    """Отправить задачу автономному разработчику DEVBOT (hive:8091)."""
-    if not ADMIN_IDS or message.from_user.id not in ADMIN_IDS:
-        await message.answer("⛔ Команда доступна только администратору.")
-        return
-
-    task = (message.text or "").removeprefix("/dev").strip()
-    if not task:
-        await message.answer(
-            "Использование: /dev <описание задачи>\n\n"
-            "Пример: /dev добавь кнопку «Повторить заказ» в бот"
-        )
-        return
-
-    status_msg = await message.answer("📤 Отправляю задачу в DEVBOT...")
-    try:
-        async with httpx.AsyncClient(timeout=10.0) as client:
-            resp = await client.post(
-                f"{DEVBOT_API_URL}/task",
-                json={"text": task},
-            )
-            resp.raise_for_status()
-            data = resp.json()
-
-        if data.get("status") == "ok":
-            await status_msg.edit_text(
-                f"✅ Задача отправлена в DEVBOT.\n\n"
-                f"📋 *{task[:200]}*\n\n"
-                f"DEVBOT уведомит тебя в Telegram когда будет готов план.",
-                parse_mode="Markdown",
-            )
-        else:
-            err = data.get("error", "неизвестная ошибка")
-            await status_msg.edit_text(f"⚠️ DEVBOT ответил: {err}")
-    except httpx.ConnectError:
-        await status_msg.edit_text(
-            "❌ DEVBOT недоступен — проверь SSH-туннель:\n"
-            "`ssh -R 8091:localhost:8091 ai-agent@185.233.200.13`",
-            parse_mode="Markdown",
-        )
-    except Exception as e:
-        logger.error("cmd_dev error: %s", e)
-        await status_msg.edit_text(f"❌ Ошибка: {e}")
-
-
-@dp.callback_query(F.data == "show_products")
-async def cb_show_products(callback: types.CallbackQuery):
-    await callback.answer()
-    await callback.message.answer(PRODUCTS_MESSAGE, reply_markup=_build_products_keyboard())
-
-
-@dp.callback_query(F.data == "show_help")
-async def cb_show_help(callback: types.CallbackQuery):
-    await callback.answer()
-    await callback.message.answer(HELP_MESSAGE, reply_markup=_build_back_to_products_keyboard())
-
-
-@dp.callback_query(F.data == "noop")
-async def cb_noop(callback: types.CallbackQuery):
-    await callback.answer()
-
-
-@dp.callback_query(F.data.startswith("ask:"))
-async def cb_ask_about_product(callback: types.CallbackQuery):
-    """Prompt user to ask a question about a specific product."""
-    try:
-        idx = int(callback.data.split(":")[1])
-        _, name, _fname, _cat = INSTRUCTIONS[idx]
-    except (ValueError, IndexError):
-        await callback.answer()
-        await callback.message.answer(ASK_MESSAGE)
-        return
-
-    await callback.answer()
-    await callback.message.answer(
-        f"Напишите свой вопрос о продукте «{name}» — я отвечу 🐝"
-    )
-
-
-def _should_respond(message: types.Message) -> bool:
-    if message.chat.type == ChatType.PRIVATE:
-        return True
-    if message.chat.id in ACTIVE_GROUP_IDS:
-        return True
-    text = message.text or ""
-    if f"@{BOT_USERNAME}" in text:
-        return True
-    if message.reply_to_message and message.reply_to_message.from_user:
-        if message.reply_to_message.from_user.id == bot.id:
-            return True
-    return False
-
-
-@dp.message(F.voice)
-async def handle_voice(message: types.Message):
-    """Respond to voice messages with a text prompt."""
-    if not _should_respond(message):
-        return
-    await message.reply(VOICE_MESSAGE)
-
-
-@dp.message(StateFilter(None))
-async def handle_question(message: types.Message, state: FSMContext):
-    """Handle user questions: search KB → generate response via Groq."""
-    if message.text and message.text.startswith("/"):
-        return
-
-    if not _should_respond(message):
-        return
-
-    query = (message.text or "").replace(f"@{BOT_USERNAME}", "").strip()
-    if len(query) < 3:
-        await message.reply("Напиши вопрос подлиннее, чтобы я мог помочь.")
-        return
-
-    # Режим «Ассистент пчеловода» — прямой LLM-диалог с CRM-контекстом
-    if ADMIN_IDS and message.from_user.id in ADMIN_IDS and message.from_user.id in _admin_mode_users:
-        await bot.send_chat_action(message.chat.id, "typing")
-        try:
-            response = await admin_chat_agent.chat(message.from_user.id, query, snapshot=_crm_snapshot)
-            try:
-                await message.reply(response, parse_mode="Markdown")
-            except Exception:
-                await message.reply(response)
-        except Exception as e:
-            logger.error("AdminChat error: %s", e)
-            await message.reply("Не удалось получить ответ. Попробуй ещё раз.")
-        return
-
-    if is_products_query(query):
-        await message.reply(PRODUCTS_MESSAGE, reply_markup=_build_products_keyboard())
-        return
-
-    logger.info(f"Question from {message.from_user.id} in {message.chat.type}: {query}")
-    await bot.send_chat_action(message.chat.id, "typing")
-
-    style = _user_styles.get(message.from_user.id)
-    user_name = message.from_user.first_name or None
-    try:
-        response, chunks = await orchestrator.route(message.from_user.id, query, style=style, user_name=user_name)
-        intent = orchestrator.get_intent(message.from_user.id)
-        logger.info(f"Intent: {intent}, chunks: {len(chunks)}")
-
-        # Маршрутизация по intent
-        if intent == "order":
-            await cmd_order(message, state)
-            return
-        if intent == "track":
-            await _handle_track(message)
-            return
-        if intent == "edit":
-            await _handle_edit(message)
-            return
-        if intent == "greeting":
-            await message.reply(response)
-            return
-
-        keyboard = _get_instruction_keyboard(chunks)
-        await message.reply(response, reply_markup=keyboard)
-
-    except Exception as e:
-        logger.error(f"Error handling question: {e}")
-        await message.reply(
-            "Извини, что-то пошло не так. Попробуй спросить ещё раз чуть позже."
-        )
-
-
-# ===========================================================================
-# Обработчики intent: track (отслеживание) и edit (редактирование)
-# ===========================================================================
-
-
-async def _handle_track(message: types.Message) -> None:
-    """Показать заказы клиента с их статусами и трек-номерами."""
-    tg_id = message.from_user.id
-    try:
-        crm = logist._crm
-        if not crm:
-            await message.reply(
-                "Для отслеживания заказа напишите Александру — "
-                "он подскажет статус и трек-номер."
-            )
-            return
-
-        orders = await crm.get_orders(client_id=None)
-        # Найти клиента по telegram_id
-        clients = await crm.get_clients()
-        client = next((c for c in clients if c.telegram_id == tg_id), None)
-        if not client:
-            await message.reply(
-                "Не нашёл ваших заказов. Если вы оформляли заказ — "
-                "напишите Александру, он поможет."
-            )
-            return
-
-        my_orders = [o for o in orders if o.client_id == client.id]
-        if not my_orders:
-            await message.reply(
-                "У вас пока нет заказов. Напишите /order чтобы оформить."
-            )
-            return
-
-        # Формат: последние 5 заказов
-        lines = ["📦 *Ваши заказы:*\n"]
-        for o in my_orders[-5:]:
-            status_emoji = {
-                "Новый": "🆕", "Подтверждён": "✅", "В сборке": "📦",
-                "Отправлен": "🚚", "Доставлен": "🎉", "Отменён": "❌",
-            }.get(o.status, "❓")
-            line = f"{status_emoji} *#{o.number}* — {o.status}"
-            if o.tracking_number:
-                line += f"\n   Трек: `{o.tracking_number}`"
-            if o.total:
-                line += f" · {o.total:.0f} ₽"
-            lines.append(line)
-
-        lines.append("\nЕсли есть вопросы — напишите Александру.")
-        await message.reply("\n".join(lines), parse_mode="Markdown")
-
-    except Exception as e:
-        logger.error("Ошибка отслеживания: %s", e)
-        await message.reply(
-            "Не удалось загрузить информацию о заказах. Попробуйте позже."
-        )
-
-
-async def _handle_edit(message: types.Message) -> None:
-    """Показать редактируемые заказы клиента."""
-    tg_id = message.from_user.id
-    editable_statuses = {"Новый", "Подтверждён", "В сборке"}
-
-    try:
-        crm = logist._crm
-        if not crm:
-            await message.reply(
-                "Для изменения заказа напишите Александру — он внесёт правки."
-            )
-            return
-
-        clients = await crm.get_clients()
-        client = next((c for c in clients if c.telegram_id == tg_id), None)
-        if not client:
-            await message.reply("Не нашёл ваших заказов.")
-            return
-
-        orders = await crm.get_orders(client_id=None)
-        my_orders = [
-            o for o in orders
-            if o.client_id == client.id and o.status in editable_statuses
-        ]
-
-        if not my_orders:
-            await message.reply(
-                "Нет заказов, которые можно изменить.\n"
-                "Заказы доступны для редактирования в статусах: "
-                "Новый, Подтверждён, В сборке."
-            )
-            return
-
-        # Показать заказы с кнопками
-        lines = ["✏️ *Заказы, доступные для изменения:*\n"]
-        buttons = []
-        for o in my_orders[-5:]:
-            total_str = f" · {o.total:.0f} ₽" if o.total else ""
-            lines.append(f"• *#{o.number}* — {o.status}{total_str}")
-            buttons.append([InlineKeyboardButton(
-                text=f"Изменить #{o.number}",
-                callback_data=f"edit_order:{o.id}",
-            )])
-
-        lines.append(
-            "\nНажмите кнопку ниже или напишите Александру — "
-            "он поможет внести изменения."
-        )
-
-        keyboard = InlineKeyboardMarkup(inline_keyboard=buttons)
-        await message.reply("\n".join(lines), parse_mode="Markdown", reply_markup=keyboard)
-
-    except Exception as e:
-        logger.error("Ошибка списка заказов для редактирования: %s", e)
-        await message.reply("Не удалось загрузить заказы. Попробуйте позже.")
-
-
-# ===========================================================================
-# Callback: редактирование заказа через Telegram
-# ===========================================================================
-
-
-@dp.callback_query(F.data.startswith("edit_order:"))
-async def cb_edit_order(callback: types.CallbackQuery):
-    """Показать опции редактирования конкретного заказа."""
-    order_id = int(callback.data.split(":")[1])
-    await callback.answer()
-
-    try:
-        crm = logist._crm
-        if not crm:
-            await callback.message.answer("CRM недоступна.")
-            return
-
-        order = await crm.get_order(order_id)
-        items = await crm.get_order_items(order_id)
-
-        lines = [f"📋 *Заказ #{order.number}*\n"]
-        if items:
-            lines.append("*Товары:*")
-            for item in items:
-                lines.append(f"  • {item.product_name or 'Товар'} × {item.quantity} = {item.total:.0f} ₽")
-        lines.append(f"\n🏠 Адрес: {order.delivery_address or '—'}")
-        lines.append(f"🚚 Доставка: {order.delivery_method or '—'}")
-        if order.delivery_cost:
-            lines.append(f"💰 Доставка: {order.delivery_cost:.0f} ₽")
-        lines.append(f"💰 *Итого: {order.total:.0f} ₽*")
-
-        lines.append(
-            "\nЧтобы внести изменения, напишите что хотите поменять, "
-            "например:\n"
-            "• «Поменяй адрес на г. Казань, ул. Мира 5»\n"
-            "• «Добавь ещё одну банку мёда»\n"
-            "\nАлександр обработает вашу просьбу."
-        )
-
-        await callback.message.answer("\n".join(lines), parse_mode="Markdown")
-
-    except Exception as e:
-        logger.error("Ошибка загрузки заказа %d: %s", order_id, e)
-        await callback.message.answer("Не удалось загрузить заказ.")
-
-
-# ===========================================================================
-# Helpers: таймаут диалога заказа
-# ===========================================================================
-
-
-async def _cancel_timeout(user_id: int) -> None:
-    """Отменить задачу таймаута для пользователя."""
-    async with _timeout_lock:
-        task = _timeout_tasks.pop(user_id, None)
-    if task and not task.done():
-        task.cancel()
-
-
-async def _timeout_dialog(user_id: int, chat_id: int, state: FSMContext) -> None:
-    """Завершить диалог заказа по таймауту."""
-    await asyncio.sleep(ORDER_TIMEOUT_SECONDS)
-    current = await state.get_state()
-    if current is not None:
-        await state.clear()
-        async with _timeout_lock:
-            _timeout_tasks.pop(user_id, None)
-        try:
-            await bot.send_message(
-                chat_id,
-                "⏰ Диалог оформления заказа завершён по таймауту (15 минут).\n"
-                "Напишите /order чтобы начать заново.",
-                reply_markup=ReplyKeyboardRemove(),
-            )
-        except Exception as e:
-            logger.warning("Не удалось отправить сообщение о таймауте: %s", e)
-
-
-async def _reset_timeout(user_id: int, chat_id: int, state: FSMContext) -> None:
-    """Сбросить таймаут — отменить старый и запустить новый."""
-    await _cancel_timeout(user_id)
-    task = asyncio.create_task(_timeout_dialog(user_id, chat_id, state))
-    async with _timeout_lock:
-        _timeout_tasks[user_id] = task
-
-
-# ===========================================================================
-# Вспомогательные клавиатуры для FSM-диалога
-# ===========================================================================
-
-
-def _delivery_keyboard(options: list[dict]) -> ReplyKeyboardMarkup:
-    """Клавиатура с вариантами доставки."""
-    return ReplyKeyboardMarkup(
-        keyboard=[[KeyboardButton(text=opt["label"])] for opt in options],
-        resize_keyboard=True,
-        one_time_keyboard=True,
-    )
-
-
-def _confirm_keyboard() -> ReplyKeyboardMarkup:
-    return ReplyKeyboardMarkup(
-        keyboard=[
-            [KeyboardButton(text="✅ Да, подтвердить")],
-            [KeyboardButton(text="❌ Нет, отменить")],
-        ],
-        resize_keyboard=True,
-        one_time_keyboard=True,
-    )
-
-
-# ===========================================================================
-# /order — начало диалога оформления заказа
-# ===========================================================================
-
-
-@dp.message(Command("order"))
-async def cmd_order(message: types.Message, state: FSMContext) -> None:
-    """Начать диалог оформления заказа."""
-    user_id = message.from_user.id
-    chat_id = message.chat.id
-
-    # Показать каталог
-    catalog_text, products = await logist.start_order(user_id)
-
-    # Сохранить каталог в FSM-контекст
-    products_data = [
-        {
-            "id": p.id,
-            "name": p.name,
-            "price": p.price,
-            "weight": p.weight,
-        }
-        for p in products
-    ]
-    await state.update_data(products=products_data, cart=[])
-    await state.set_state(OrderFSM.choosing_product)
-
-    await _reset_timeout(user_id, chat_id, state)
-    await message.answer(catalog_text, parse_mode="Markdown")
-
-
-# ===========================================================================
-# /cancel — отмена диалога на любом шаге
-# ===========================================================================
-
-
-@dp.message(Command("cancel"), StateFilter(OrderFSM))
-async def cmd_cancel_order(message: types.Message, state: FSMContext) -> None:
-    """Отменить диалог оформления заказа."""
-    await _cancel_timeout(message.from_user.id)
-    await state.clear()
-    await message.answer(
-        "Заказ отменён. Напишите /order чтобы начать заново.",
-        reply_markup=ReplyKeyboardRemove(),
-    )
-
-
-# ===========================================================================
-# Шаг 1: Выбор товаров
-# ===========================================================================
-
-
-@dp.message(OrderFSM.choosing_product)
-async def fsm_choose_product(message: types.Message, state: FSMContext) -> None:
-    """Обработать выбор товаров из каталога."""
-    user_id = message.from_user.id
-    data = await state.get_data()
-    products = data.get("products", [])
-
-    cart, error = logist.parse_product_selection(message.text or "", products)
-    if error:
-        await message.answer(error)
-        return
-
-    # Загрузить данные постоянного клиента
-    existing_client = await logist.get_existing_client(user_id)
-
-    await state.update_data(cart=cart)
-
-    # Сохранить телефон для предзаполнения на шаге 3
-    if existing_client and existing_client.phone:
-        await state.update_data(_existing_client_phone=existing_client.phone)
-
-    prefill = ""
-    if existing_client and existing_client.full_name:
-        name = existing_client.full_name
-        # Не предлагать автоподстановку если имя = "Telegram XXXXX"
-        if not name.startswith("Telegram "):
-            prefill = f"\nПодсказка: в прошлый раз вы представились как *{name}*.\nПросто отправьте *да* или введите другое."
-            await state.update_data(prefill_name=name)
-
-    await state.set_state(OrderFSM.entering_name)
-    await _reset_timeout(user_id, message.chat.id, state)
-    await message.answer(
-        f"Отлично! Введите ваше *ФИО* (Фамилия Имя Отчество).{prefill}",
-        parse_mode="Markdown",
-    )
-
-
-# ===========================================================================
-# Шаг 2: Ввод ФИО
-# ===========================================================================
-
-
-@dp.message(OrderFSM.entering_name)
-async def fsm_enter_name(message: types.Message, state: FSMContext) -> None:
-    """Обработать ввод ФИО."""
-    name = (message.text or "").strip()
-    if len(name) < 3:
-        await message.answer("Пожалуйста, введите полное ФИО (минимум 3 символа).")
-        return
-
-    data = await state.get_data()
-    # Если пользователь нажал «Enter» без текста — использовать предзаполненное
-    if name.lower() in ("да", "+", "ok", "ок") and data.get("prefill_name"):
-        name = data["prefill_name"]
-
-    await state.update_data(full_name=name)
-
-    # Предзаполнение телефона из данных существующего клиента
-    data = await state.get_data()
-    phone_prefill = ""
-    existing_client = data.get("_existing_client_phone")
-    if existing_client:
-        phone_prefill = (
-            f"\nПодсказка: ваш прошлый номер — *{format_phone(existing_client)}*.\n"
-            "Отправьте *да* или введите другой."
-        )
-        await state.update_data(prefill_phone=existing_client)
-
-    await state.set_state(OrderFSM.entering_phone)
-    await _reset_timeout(message.from_user.id, message.chat.id, state)
-    await message.answer(
-        f"📞 Введите ваш *номер телефона*:{phone_prefill}",
-        parse_mode="Markdown",
-    )
-
-
-# ===========================================================================
-# Шаг 3: Ввод телефона
-# ===========================================================================
-
-
-@dp.message(OrderFSM.entering_phone)
-async def fsm_enter_phone(message: types.Message, state: FSMContext) -> None:
-    """Обработать ввод номера телефона."""
-    phone_raw = (message.text or "").strip()
-    data = await state.get_data()
-
-    # Если пользователь подтвердил предзаполненный номер
-    if phone_raw.lower() in ("да", "+", "ok", "ок") and data.get("prefill_phone"):
-        phone = data["prefill_phone"]
-    else:
-        phone, error = validate_phone(phone_raw)
-        if phone is None:
-            await message.answer(error)
-            return
-
-    existing_client = None
-    if data.get("prefill_address") is None:
-        existing_client = await logist.get_existing_client(message.from_user.id)
-
-    await state.update_data(phone=phone)
-
-    prefill = ""
-    if existing_client and existing_client.address:
-        prefill = (
-            f"\nПодсказка: ваш прошлый адрес — *{existing_client.address}*.\n"
-            "Отправьте его или введите новый."
-        )
-        await state.update_data(prefill_address=existing_client.address)
-
-    await state.set_state(OrderFSM.entering_address)
-    await _reset_timeout(message.from_user.id, message.chat.id, state)
-    await message.answer(
-        f"🏠 Введите *адрес доставки* (город, улица, дом, квартира).{prefill}",
-        parse_mode="Markdown",
-    )
-
-
-# ===========================================================================
-# Шаг 4: Ввод адреса
-# ===========================================================================
-
-
-@dp.message(OrderFSM.entering_address)
-async def fsm_enter_address(message: types.Message, state: FSMContext) -> None:
-    """Обработать ввод адреса доставки."""
-    address = (message.text or "").strip()
-    if len(address) < 5:
-        await message.answer("Введите полный адрес (минимум 5 символов).")
-        return
-
-    data = await state.get_data()
-    # Если ввели подтверждение — использовать предзаполненный адрес
-    if address.lower() in ("да", "+", "ok", "ок") and data.get("prefill_address"):
-        address = data["prefill_address"]
-
-    cart = data.get("cart", [])
-    options = await logist.get_delivery_options(cart, address=address)
-
-    await state.update_data(address=address, delivery_options=options)
-    await state.set_state(OrderFSM.choosing_delivery)
-    await _reset_timeout(message.from_user.id, message.chat.id, state)
-
-    keyboard = _delivery_keyboard(options)
-    options_text = "\n".join(f"  {opt['label']}" for opt in options)
-    await message.answer(
-        f"🚚 Выберите *способ доставки*:\n\n{options_text}",
-        parse_mode="Markdown",
-        reply_markup=keyboard,
-    )
-
-
-# ===========================================================================
-# Шаг 5: Выбор способа доставки
-# ===========================================================================
-
-
-@dp.message(OrderFSM.choosing_delivery)
-async def fsm_choose_delivery(message: types.Message, state: FSMContext) -> None:
-    """Обработать выбор способа доставки."""
-    user_input = (message.text or "").strip()
-    data = await state.get_data()
-    options = data.get("delivery_options", [])
-
-    # Сопоставить ввод пользователя с вариантами
-    chosen = None
-    chosen_cost = 0.0
-    for opt in options:
-        if opt["method"].lower() in user_input.lower() or opt["label"] in user_input:
-            chosen = opt["method"]
-            chosen_cost = opt["cost"]
-            break
-
-    if not chosen:
-        labels = "\n".join(f"  • {opt['label']}" for opt in options)
-        await message.answer(
-            f"Пожалуйста, выберите один из вариантов:\n{labels}",
-            reply_markup=_delivery_keyboard(options),
-        )
-        return
-
-    cart = data.get("cart", [])
-    full_name = data.get("full_name", "")
-    phone = data.get("phone", "")
-    address = data.get("address", "")
-
-    summary = format_order_summary(
-        cart=cart,
-        full_name=full_name,
-        phone=phone,
-        address=address,
-        delivery=chosen,
-        delivery_cost=chosen_cost,
-    )
-
-    await state.update_data(delivery=chosen, delivery_cost=chosen_cost)
-    await state.set_state(OrderFSM.confirming_order)
-    await _reset_timeout(message.from_user.id, message.chat.id, state)
-
-    await message.answer(
-        summary,
-        parse_mode="Markdown",
-        reply_markup=_confirm_keyboard(),
-    )
-
-
-# ===========================================================================
-# Шаг 6: Подтверждение заказа
-# ===========================================================================
-
-
-@dp.message(OrderFSM.confirming_order)
-async def fsm_confirm_order(message: types.Message, state: FSMContext) -> None:
-    """Обработать подтверждение или отмену заказа."""
-    answer = (message.text or "").strip().lower()
-
-    if answer in ("да", "✅ да, подтвердить", "yes", "подтвердить", "+"):
-        await state.set_state(OrderFSM.creating_order)
-        await _do_create_order(message, state)
-    elif answer in ("нет", "❌ нет, отменить", "no", "отменить", "-"):
-        await _cancel_timeout(message.from_user.id)
-        await state.clear()
-        await message.answer(
-            "Заказ отменён. Напишите /order чтобы начать заново.",
-            reply_markup=ReplyKeyboardRemove(),
-        )
-    else:
-        await message.answer(
-            "Пожалуйста, отправьте *да* для подтверждения или *нет* для отмены.",
-            parse_mode="Markdown",
-            reply_markup=_confirm_keyboard(),
-        )
-
-
-# ===========================================================================
-# Шаг 7: Создание заказа
-# ===========================================================================
-
-
-async def _do_create_order(message: types.Message, state: FSMContext) -> None:
-    """Создать заказ в Integram и уведомить пчеловода."""
-    await _cancel_timeout(message.from_user.id)
-    data = await state.get_data()
-    await state.clear()
-
-    await bot.send_chat_action(message.chat.id, "typing")
-
-    success, response_text = await logist.create_order(
-        telegram_id=message.from_user.id,
-        full_name=data.get("full_name", ""),
-        phone=data.get("phone", ""),
-        address=data.get("address", ""),
-        delivery=data.get("delivery", ""),
-        delivery_cost=data.get("delivery_cost", 0.0),
-        cart=data.get("cart", []),
-        telegram_username=message.from_user.username,
-    )
-
-    await message.answer(
-        response_text,
-        parse_mode="Markdown",
-        reply_markup=ReplyKeyboardRemove(),
-    )
-
-    if success:
-        # Уведомить пчеловода
-        cart = data.get("cart", [])
-        delivery_cost = data.get("delivery_cost", 0.0)
-        items_total = sum(i["qty"] * i["unit_price"] for i in cart)
-        total = items_total + delivery_cost
-        items_str = "\n".join(f"  • {i['name']} × {i['qty']}" for i in cart)
-        beekeeper_msg = (
-            f"👤 {data.get('full_name', '')}\n"
-            f"📞 {data.get('phone', '')}\n"
-            f"🏠 {data.get('address', '')}\n"
-            f"🚚 {data.get('delivery', '')} — {delivery_cost:.0f} ₽\n\n"
-            f"*Товары:*\n{items_str}\n\n"
-            f"💰 Итого: {total:.0f} ₽"
-        )
-        await logist.notify_beekeeper(bot, beekeeper_msg)
-
-
-@dp.callback_query(F.data.startswith("doc:"))
-async def send_instruction_pdf(callback: types.CallbackQuery):
-    """Send the instruction PDF when user taps the button."""
-    try:
-        idx = int(callback.data.split(":")[1])
-        _, name, filename, _cat = INSTRUCTIONS[idx]
-    except (ValueError, IndexError):
-        await callback.answer("Инструкция не найдена.", show_alert=True)
-        return
-
-    pdf_path = PDFS_DIR / filename
-    if not pdf_path.exists():
-        await callback.answer("Файл не найден на сервере.", show_alert=True)
-        return
-
-    await callback.answer()
-    await callback.message.answer_document(
-        document=FSInputFile(str(pdf_path), filename=filename),
-        caption=f"📄 {name}",
-        reply_markup=InlineKeyboardMarkup(inline_keyboard=[[
-            InlineKeyboardButton(text=f"❓ Задать вопрос о {name}", callback_data=f"ask:{idx}"),
-        ], [
-            InlineKeyboardButton(text="📦 Все продукты", callback_data="show_products"),
-        ]]),
-    )
+# bot_admin_router — admin-команды из bot.py (stats, yt_*, faq, dev, admin-mode)
+dp.include_router(bot_admin_router)
+# inspect_router — InspectFSM (до OrderFSM чтобы перехватить состояния)
+dp.include_router(inspect_router)
+# fsm_order_router — OrderFSM + callbacks продуктов/инструкций
+dp.include_router(fsm_order_router)
+# worker_router — очередь сборки работников склада
+dp.include_router(worker_router)
+# user_router — ПОСЛЕДНИМ (StateFilter(None) — ловит все остальные сообщения)
+dp.include_router(user_router)
 
 
 async def _alert(text: str) -> None:
@@ -1769,31 +70,44 @@ async def on_startup(**_kwargs) -> None:
 async def main():
     global bot
     setup_logging()
-    # Инициализировать SOCKS5-прокси здесь, внутри event loop
+
     if TG_SOCKS_PROXY:
         from aiogram.client.session.aiohttp import AiohttpSession
         bot = Bot(token=TELEGRAM_BOT_TOKEN, session=AiohttpSession(proxy=TG_SOCKS_PROXY))
         logger.info("Telegram via SOCKS5 proxy: %s", TG_SOCKS_PROXY)
+
     logger.info("Starting BEEBOT...")
+
+    # --- Загрузка базы знаний ---
+    orchestrator = Orchestrator()
+    inspector = InspectorAgent()
+    logist = LogistAgent(beekeeper_chat_id=BEEKEEPER_CHAT_ID)
+    analyst = AnalystAgent(
+        groq_client=orchestrator._groq,
+        groq_model=orchestrator._model,
+    )
+    admin_chat_agent = AdminChatAgent(
+        groq_client=orchestrator._groq,
+        model=orchestrator._model,
+    )
+
     try:
         orchestrator.load_kb()
-        inspector.kb = orchestrator._beebot.kb  # разделяем KB с beebot
+        inspector.kb = orchestrator._beebot.kb
         kb = orchestrator._beebot.kb
-        logger.info(f"Knowledge base loaded: {len(kb.chunks)} chunks")
+        logger.info("Knowledge base loaded: %d chunks", len(kb.chunks))
     except FileNotFoundError:
         logger.error("Knowledge base not found! Run `python -m src.build_kb` first.")
         return
 
-    # --- Integram CRM: общий клиент для всех агентов ---
+    # --- Integram CRM ---
     integram_client: Optional[IntegramClient] = None
     if app_config.INTEGRAM_URL and app_config.INTEGRAM_LOGIN:
         try:
             integram_client = IntegramClient()
             await integram_client.authenticate()
-            # Передать CRM-клиент всем агентам
             logist._crm = integram_client
             analyst._crm = integram_client
-            # Динамический keyword-буст из CRM
             try:
                 products = await integram_client.get_products()
                 names = [p.name for p in products if p.name]
@@ -1808,37 +122,38 @@ async def main():
     else:
         logger.info("Integram CRM не настроена — агенты работают без CRM.")
 
-    # Загрузить онтологию Симптомы→Показания из Integram
+    # --- Онтология ---
     try:
         await orchestrator.load_ontology()
     except Exception as _e:
         logger.warning("Онтология недоступна (продолжаем без неё): %s", _e)
 
-    # Инициализировать AdminChatAgent с CRM-клиентом
     admin_chat_agent.set_crm(integram_client)
-
-    # Инициализировать админ-модуль (один раз, с CRM если доступна)
     setup_admin(bot, crm=integram_client, kb=kb, memory=orchestrator._memory)
+
+    # --- Инициализация роутеров ---
+    setup_inspect(inspector)
+    setup_fsm_order(logist, bot)
+    setup_bot_admin(analyst, orchestrator, admin_chat_agent, inspector, bot)
+    setup_user(orchestrator, admin_chat_agent, logist)
 
     # --- Режим работника склада ---
     if integram_client:
-        global _worker_crm
-        _worker_crm = integram_client
+        setup_worker(integram_client, bot)
         from src.notifications import Notifier
         import src.notifications as _notif_module
         _notif_module._worker_notifier = Notifier(bot)
         if WORKER_CHAT_IDS:
             logger.info("Режим работника склада включён (%d работников).", len(WORKER_CHAT_IDS))
 
-    # --- CRM Snapshot: кэш заказов с позициями ---
+    # --- CRM Snapshot ---
     if integram_client:
-        global _crm_snapshot
         from src.crm_snapshot import CrmSnapshot
-        _crm_snapshot = CrmSnapshot(integram_client)
-        asyncio.create_task(_crm_snapshot.run())
-        logger.info("CRM snapshot запущен (интервал %d сек).", _crm_snapshot._refresh_interval)
+        _state._crm_snapshot = CrmSnapshot(integram_client)
+        asyncio.create_task(_state._crm_snapshot.run())
+        logger.info("CRM snapshot запущен (интервал %d сек).", _state._crm_snapshot._refresh_interval)
 
-    # --- Авто-трекинг: фоновая проверка статуса отправлений ---
+    # --- Авто-трекинг ---
     order_tracker: Optional[OrderTracker] = None
     if integram_client:
         from src.web.notifications import notify_client_status_change
@@ -1849,7 +164,7 @@ async def main():
         asyncio.create_task(order_tracker.run())
         logger.info("Авто-трекинг отправлений запущен.")
 
-    # --- UDS Poller: фоновая синхронизация заказов из UDS ---
+    # --- UDS Poller ---
     uds_poller: Optional[UDSPoller] = None
     uds_client: Optional[UDSClient] = None
 
@@ -1870,7 +185,7 @@ async def main():
         else:
             logger.warning("UDS Poller пропущен — Integram CRM не подключена.")
     else:
-        logger.info("UDS не настроен (UDS_API_KEY/UDS_COMPANY_ID не заданы) — поллер пропущен.")
+        logger.info("UDS не настроен — поллер пропущен.")
 
     logger.info("Bot is running!")
     _crashed = False
@@ -1884,7 +199,6 @@ async def main():
     finally:
         if not _crashed:
             await _alert("🔴 BEEBOT остановлен")
-        # Graceful shutdown: остановить трекер, UDS poller и закрыть клиенты
         if order_tracker:
             order_tracker.stop()
         if uds_poller:
